requirements.txt
asgiref==3.4.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==4.3.0
Django==3.2.9
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-filter==21.1
django-mpesa==2.0.8
django-nose==1.4.7
django-reversion==2.0.13
django-uni-form==0.9.0
djangorestframework==3.12.4
djangorestframework-gis==0.11.2
djappsettings==0.4.0
djtables==0.1.2
enum34==1.1.6
idna==3.3
importlib-metadata==2.1.2
ipython==5.6.0
ipython-genutils==0.2.0
nose==1.3.7
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==8.3.2
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-slugify==1.2.5
pytz==2018.3
RapidSMS==0.9.6a0
rapidsms-xforms==0.3.0
requests==2.26.0
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
soupsieve==2.3.1
sqlparse==0.4.2
traitlets==4.3.2
typing-extensions==3.10.0.2
Unidecode==1.0.22
urllib3==1.26.7
wcwidth==0.1.7
zipp==3.6.0
